Fixed client-js not reconnecting on certain circumstances
This is a special case for context deadline exceed that can happen within `graphql-go`
library directly. In the library, there is a maximal time for of subscription execution to
happen. We customize this to 10 seconds in `schema.go#parseSchema` but it still can
happen under abnormal network condition that the timeout kicks in.

When the timeout happen through the `context.Err()`, the error is then wrapped manually
in a `gqerrors.QueryError` struct and the `context.Err()` string is put straight as-is
in the `Message` field. We need to forward that as a `terminal` error otherwise the client
will not re-connect automatically.

However, to make things harder, the context that has the timeout value is not under our control
so we cannot use our subscription context and check for its `Err()` value. So we need to check the
actual error message value manually.

Hackylish, if the message in the error is `context.DeadlineExceeded.Error()`, we return a custom
`gqerrors.QueryError` back as a terminal error.

@see https://github.com/graph-gophers/graphql-go/blob/dae41bde9ef91c12e78863f0299348612f2d6214/internal/exec/subscribe.go#L115 (Subscription timeout context)
@see https://github.com/graph-gophers/graphql-go/blob/dae41bde9ef91c12e78863f0299348612f2d6214/internal/exec/subscribe.go#L65 (Subscription context error wrapping)